Finish the solution for problem 'Minimum Path Sum'

+++ src/Unique Paths.cpp
@@ -3,7 +3,7 @@
 class Solution {
 public:
   int uniquePaths(int m, int n) {
-    return Factorial(std::max(m - 1, n - 1) + 1, m - 1 + n - 1) / Factorial(1, std::min(m - 1, n - 1));
+    return static_cast<int>(Factorial(std::max(m - 1, n - 1) + 1, m - 1 + n - 1) / Factorial(1, std::min(m - 1, n - 1)));
   }
 
   long long Factorial(long long start, long long end) {
